Speed up by 10%: use optimal threadsPerThreadgroup values for each layer.
Implement Net.ProfileLayer and add a corresponding unit test (rather, benchmark),
which find the optimal values and print them to the console.

While overal improvement is not so great, some particular layers are up to 30% faster.

Note: this does not include the preprocess layer and the very top layers,
which are currently hardcoded. Probably, it's worth looking at.